requirements.txt
-i https://pypi.org/simple
apscheduler==3.9.1
beautifulsoup4==4.11.1
certifi==2022.6.15; python_full_version >= '3.6.0'
charset-normalizer==2.1.1; python_full_version >= '3.6.0'
cooltables==0.0.3
icalendar==4.1.0
idna==3.3; python_version >= '3.5'
lectio.py==0.1.0
lxml==4.9.1
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
requests==2.28.1
setuptools==65.1.1; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.3.2.post1; python_full_version >= '3.6.0'
tzdata==2022.2; python_full_version >= '3.6.0'
tzlocal==4.2; python_full_version >= '3.6.0'
urllib3==1.26.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
